+++ mobile/android/android-components/docs/changelog.md
@@ -5,6 +5,8 @@ permalink: /changelog/
 ---
 
 # 151.0 (In Development)
+* **browser-engine-gecko**
+  * 🆕 New Engine Settings API - `useContentBlockingDatabase` that allows controlling whether the content blocking database is enabled. [Bug 2030082](https://bugzilla.mozilla.org/show_bug.cgi?id=2030082)
 
 # 150.0
 * **browser-engine-gecko**